+++ src/Constants.php
@@ -4,5 +4,5 @@ namespace Microsoft\Kiota\Authentication;
 
 class Constants
 {
-    public const VERSION = "1.0.2";
+    public const VERSION = "1.1.0";
 }